Add files via upload
Since the file is too large, so I can only upload seperately. The Basic.R is the preprocessing and machine learning part for finding the algorithms. The server.r, ui.R and trained_model.RData are together composed the application. They need to be put into the same folder , then it can run.